add entry trait
