+++ src/dedup.rs
@@ -1,44 +1,10 @@
-use std::cmp::Ordering;
 use std::collections::HashMap;
-use std::path::Path;
+use entry::Entry;
 use file::FileHash;
-use walkdir::DirEntry;
 
-#[derive(Debug)]
-pub struct SortableDirEntry(DirEntry);
-
-impl SortableDirEntry {
-    pub fn new(entry: DirEntry) -> SortableDirEntry {
-        SortableDirEntry(entry)
-    }
-    
-    pub fn path(&self) -> &Path {
-        self.0.path()
-    }
-}
-
-impl PartialEq for SortableDirEntry {
-    fn eq(&self, other: &Self) -> bool {
-        self.path().eq(other.path())
-    }
-}
-
-impl Eq for SortableDirEntry { }
-
-impl PartialOrd for SortableDirEntry {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.path().partial_cmp(other.path())
-    }
-}
-
-impl Ord for SortableDirEntry {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.path().cmp(other.path())
-    }
-}
-
-pub fn group_by_size<I>(paths: I) -> Vec<Vec<SortableDirEntry>>
-    where I: Iterator<Item = (u64, SortableDirEntry)>
+pub fn group_by_size<E, I>(paths: I) -> Vec<Vec<E>>
+    where E: Entry,
+          I: Iterator<Item = (u64, E)>
 {
     let groups = paths.fold(HashMap::new(), |mut map, (len, entry)| {
         map.entry(len).or_insert_with(Vec::new).push(entry);
@@ -51,8 +17,9 @@ pub fn group_by_size<I>(paths: I) -> Vec<Vec<SortableDirEntry>>
         .collect()
 }
 
-pub fn group_by_hash<I>(paths: I, verbose: bool) -> Vec<(Vec<u8>, Vec<SortableDirEntry>)>
-    where I: Iterator<Item = SortableDirEntry>
+pub fn group_by_hash<E, I>(paths: I, verbose: bool) -> Vec<(Vec<u8>, Vec<E>)>
+    where E: Entry + Ord,
+          I: Iterator<Item = E>
 {
     let groups = paths.filter_map(|entry| {
             if verbose {
